Move bean registration to separate classes
